requirements.txt
asgiref>=3.2.3
reverse-geocoder==1.5.1
certifi>=2019.11.28
chardet>=3.0.4
dj-database-url>=0.5.0
Django>=3.0.2
django-cors-headers>=3.2.1
django-postgres-copy==2.3.1
python-decouple==3.1
django-heroku>=0.3.1
django-allauth>=0.41.0
future>=0.18.2
geocoder>=1.38.1
gunicorn>=20.0.4
python-dateutil>=2.8.1
pytz>=2019.3
requests>=2.22.0
six>=1.14.0
sqlparse>=0.3.0
whitenoise>=5.0.1
django-reset-migrations==0.4.0